feat(profile): merge settings into profile page
Combine the Profile and Settings screens into a single Profile
destination. The merged page now hosts the appearance controls
(theme mode + color scheme) and the sign-out button alongside the
existing account and about sections.

- Move theme/color/sign-out widgets into profile_widgets.dart
- Remove the /settings shell branch, route, and nav destination
- Delete the now-empty settings feature
- Regenerate router.g.dart
