--- lib/features/settings/presentation/screens/settings_screen.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../widgets/settings_widgets.dart';
-
-class SettingsScreen extends StatelessWidget {
-  const SettingsScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) => const SettingsBody();
-}
